Lab Refactor
I did the following refactoring:

In LonelyTwitterActivity.java, mTweets and adapter is changed to be local variable
The name of the Tweet.java class is be changed to AbstractTweet.java, so that it's easy to know it's an abstract class.
In Tweet.java or AbstractTweet.java, the upper limit of a tweet can be, it was a constant 140, but I created a new private variable called limit to store the upper limit. So that it's easier to modify the upper limit.
In LonelyTwitterActivity.java, numImportant is set to be private
In LonelyTwitterActivity.java, REQUEST_CAPTURING_IMAGE is changed to be private static final.